+++ playa/ascii85.py
@@ -4,6 +4,9 @@ import re
 from base64 import a85decode
 from binascii import unhexlify
 
+start_re = re.compile(rb"^\s*<?\s*~\s*")
+end_re = re.compile(rb"\s*~\s*>?\s*$")
+
 
 def ascii85decode(data: bytes) -> bytes:
     """In ASCII85 encoding, every four bytes are encoded with five ASCII
@@ -11,15 +14,17 @@ def ascii85decode(data: bytes) -> bytes:
     When the length of the original bytes is not a multiple of 4, a special
     rule is used for round up.
 
-    The Adobe's ASCII85 implementation is slightly different from
-    its original in handling the last characters.
-
+    Adobe's ASCII85 implementation expects the input to be surrounded
+    by `b"<~"` and `b"~>"`.  We can't reliably expect this to be the
+    case, and often there are off-by-one errors in xref tables or
+    stream lengths which mean we only see `~`.  Worse yet, `<` and `>`
+    are ASCII85 digits (yes, another genius move on Adobe's part), so
+    we can't just strip them out too.  We settle on a compromise where
+    we strip leading `<~` or `~` and trailing `~` or `~>`.
     """
-    data = data.strip()
-    try:
-        return a85decode(data, adobe=True)
-    except ValueError:
-        return a85decode(data)
+    data = start_re.sub(b"", data)
+    data = end_re.sub(b"", data)
+    return a85decode(data)
 
 
 bws_re = re.compile(rb"\s")