+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## PLAYA 0.4.2: Unreleased
 - Correct `fontsize` and `scaling` in text state
+- Correct `ValueError` on incorrect stream lengths for ASCII85 data
 - Correct implicit font encodings for Type1 fonts
 - More fine-grained error handling in font initialization